third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/break-on-set-timeout-with-syntax-error-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected identifier
Tests that scheduled pause is cleared after processing event with empty handler.

Call stack:
    0) executeSomeCode (break-on-set-timeout-with-syntax-error.html:15)
    1)  (:1)
